@@ -1,4 +1,4 @@
-class AddGatherersTrackingToPlayers < ActiveRecord::Migration
+class AddGatherersTrackingToPlayers < ActiveRecord::Migration[4.2]
   def change
     add_column :players, :mining_ehp_start, :float
     add_column :players, :mining_ehp_end, :float
