@@ -1,4 +1,4 @@
-class AddCombatLevelToPlayers < ActiveRecord::Migration
+class AddCombatLevelToPlayers < ActiveRecord::Migration[4.2]
   def change
     add_column :players, :combat_lvl, :integer
   end
