@@ -1,4 +1,4 @@
-class AddTimeToMaxToPlayers < ActiveRecord::Migration
+class AddTimeToMaxToPlayers < ActiveRecord::Migration[4.2]
   def change
     add_column :players, :ttm_lvl, :float
     add_column :players, :ttm_xp, :float
